Tweak bit size of fast float reduction loop
This loop chops off bits from the end of the mantissa to see if it
makes a difference to the value.  It adds "1111110" to the last
digits of the mantissa, then uses the floor function to see if that
made a difference.  The last bit must be zero due to rounding, but it
is possible that rounding rules with some ieee754 settings or on other
float platforms will be different and this loop would not work for
them.