Add more float encoding tests
Adding tests of small floats and negatives, testing that the use of
POSIX::floor is correct as opposed to, say, POSIX::trunc to round
towards 0 (which is what logically must happen, but 'abs' has already
made sure that the sign bit is clear).  Also add some tests near the
limit of the encoding precision - it's clear that quite a few bits are
wasted when encoding using ieee math directly, but the C solutions
which can extract bit fields directly should not suffer from this
problem.