Git::DB::Schema: a theory of self-describing types
The idea is that the mapping mechanism should be able to take the
class definitions in the .pm files (::Schema, ::Class, ::Attr and
::Key), and map them to the tables described in the Tables.pod file.
The Postgres.pod file lists the concrete tables in the equivalent
structures in a modern-ish postgres, and there is a column for how we
will map them.  The Tables.pod file also lists important missing
features in our mapping.

This set should hopefully be the minimum required to bootstrap the
system!