Tidy up paragraph, it didn't make sense before.
Now it might make more sense to some, and be a complete epiphany for
others.

(₁₂√2)⁷, a 5th, is damn close to 3/2, the first improper fraction, and
my theory is that that is why it sounds so awesome.

Also of relevance: for all of your other encoding requirements, there
is Unicode.