Come up with a solution for unsorted tree objects
It occurred to me as I was writing code to scan the trees that because
of the filename rules, the entries would not be coming back from the
tree in a useful enough order.  This means that all of the entries in
any given tree would have to be decoded just to look up any entry.

This might not take very long, and could be limited by making the
individual node size smaller, eg just 20 entries per node or something
like that.  It just seemed to be losing algorithmic ground too quickly
for my liking...

Anyway, this solution of a surrogate sort identifier occurred to me.
Perhaps it's a premature optimization, but it doesn't seem to be too
ugly and is a useful assumption for the tree format code to be able to
make.  I'll make it optional and see if it proves popular..