'value types' definitions.
This table, here shown as a plain CSV input, connects a simple
type name and a set of formats with a number of well–known
functions, which handle conversion from the column format to
memory and back, as well as converting to a string.  Used when
opening a schema to check that the value type semantics are known.