Add concrete 'unique' keys index type
Unique indexes constrain the values in the tables they apply to,
to be unique.  They can be the target of a foreign key constraint
or just a data integrity assertion.

Unique is a sub–type of key, but adds no attributes.  For
polymorphic select to work (eg, selecting all from key and getting
*specialized* rows back), only one sub-type of a tuple is allowed
to add no attributes.

Add a bunch of other unique constraints on the schema.